requirements.txt
mediapipe
cvzone
pyserial
